scripts/backfill/requirements.txt
beautifulsoup4
google-cloud-pubsub
requests
tqdm
youtube_transcript_api
